Extended our first view chapter #2
I’ve added a second view to the first chapter so the reader can be
quickly moving into the mindset of this ToDo app. I think it’ll be
easier to get something that looks like it’s going to work, then show
how it can be improved, with new features added, by using Marionette’s
features.