Quick template overview
I’ve written up an overview of the template system in Marionette. I’ve
only covered Underscore as that’s the default template engine in the
framework.

I’m open to others adding in their own - just make sure to include a
how-to either here or in the “Installing Marionette” section.